Python-1/requirements.txt
numpy
Pillow
matplotlib.pyplot
